+++ src/domain/flow-system/services/src/flow/active_configs_state.rs
@@ -19,6 +19,7 @@ pub(crate) struct ActiveConfigsState {
     dataset_schedules: HashMap<FlowKeyDataset, Schedule>,
     system_schedules: HashMap<SystemFlowType, Schedule>,
     dataset_batching_rules: HashMap<FlowKeyDataset, BatchingRule>,
+    dataset_reset_rules: HashMap<FlowKeyDataset, ResetRule>,
     dataset_compaction_rules: HashMap<FlowKeyDataset, CompactionRule>,
 }
 
@@ -32,6 +33,9 @@ impl ActiveConfigsState {
         match rule {
             FlowConfigurationRule::Schedule(schedule) => {
                 self.dataset_schedules.insert(key, schedule);
+            }
+            FlowConfigurationRule::ResetRule(reset) => {
+                self.dataset_reset_rules.insert(key, reset);
             }
             FlowConfigurationRule::BatchingRule(batching) => {
                 self.dataset_batching_rules.insert(key, batching);
@@ -102,6 +106,16 @@ impl ActiveConfigsState {
             .copied()
     }
 
+    pub fn try_get_dataset_reset_rule(
+        &self,
+        dataset_id: &DatasetID,
+        flow_type: DatasetFlowType,
+    ) -> Option<ResetRule> {
+        self.dataset_reset_rules
+            .get(BorrowedFlowKeyDataset::new(dataset_id, flow_type).as_trait())
+            .cloned()
+    }
+
     pub fn try_get_config_snapshot_by_key(
         &self,
         flow_key: &FlowKey,
@@ -120,6 +134,12 @@ impl ActiveConfigsState {
                 DatasetFlowType::Ingest => self
                     .try_get_flow_schedule(flow_key)
                     .map(FlowConfigurationSnapshot::Schedule),
+                DatasetFlowType::Reset => self
+                    .try_get_dataset_reset_rule(
+                        &dataset_flow_key.dataset_id,
+                        dataset_flow_key.flow_type,
+                    )
+                    .map(FlowConfigurationSnapshot::Reset),
                 DatasetFlowType::HardCompaction => self
                     .try_get_dataset_compaction_rule(
                         &dataset_flow_key.dataset_id,