+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- New `Reset` flow in GQL Api which can be triggered manually for `Root` and `Derivative` datasets
+
 ## [0.193.1] - 2024-08-09
 ### Fixed
 - Panic for `EXECUTE_TRANSFORM` flow without dataset env vars enabled feature